% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{availTCGA}
\alias{availTCGA}
\title{Get or Check TCGA Available ProjectID, DataType and FileType}
\usage{
availTCGA(which = c("all", "ProjectID", "DataType", "FileType"))
}
\arguments{
\item{which}{a character of \code{c("All", "ProjectID", "DataType", "FileType")}}
}
\description{
Get or Check TCGA Available ProjectID, DataType and FileType
}
\examples{
\donttest{
availTCGA("all")
}
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
