% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaHub-class.R
\name{cohorts}
\alias{cohorts}
\title{Get cohorts of XenaHub object}
\usage{
cohorts(x)
}
\arguments{
\item{x}{a \link{XenaHub} object}
}
\value{
a character vector contains cohorts
}
\description{
Get cohorts of XenaHub object
}
\examples{
xe = XenaGenerate(subset = XenaHostNames == "TCGA"); cohorts(xe)
}
