% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaHub-class.R
\name{datasets}
\alias{datasets}
\title{Get datasets of XenaHub object}
\usage{
datasets(x)
}
\arguments{
\item{x}{a \code{XenaHub} object}
}
\value{
a character vector contains datasets
}
\description{
Get datasets of XenaHub object
}
\examples{
xe = XenaHub(hostName="TCGA"); datasets(xe)
}
