% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{getTCGAdata}
\alias{getTCGAdata}
\title{Get TCGA Common Data Sets by Project ID and Property}
\usage{
getTCGAdata(project = NULL, clinical = TRUE, mRNASeq = FALSE,
  mRNAArray = FALSE, mRNASeqType = c("normalized", "pancan normalized",
  "percentile"), miRNASeq = FALSE, miRNAArray = FALSE,
  RPPAArray = FALSE, ReplicateBaseNormalization = FALSE,
  Methylation = FALSE, MethylationType = c("27K", "450K"),
  GeneMutation = FALSE, SomaticMutation = FALSE,
  GisticCopyNumber = FALSE, Gistic2Threshold = TRUE,
  CopyNumberSegment = FALSE, RemoveGermlineCNV = TRUE,
  download = FALSE, forceDownload = FALSE)
}
\description{
This is the most useful function for user to download common TCGA datasets,
it is similar to \code{getFirehoseData} function in \code{RTCGAToolbox} package. If you are
not familiar with data structure of TCGA on Xena, please visit
<https://xenabrowser.net/datapages/?host=https%3A%2F%2Ftcga.xenahubs.net&removeHub=https%3A%2F%2Fxena.treehouse.gi.ucsc.edu> and
select one tumor type (or what you want to download).
}
