% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaHub-class.R
\name{hosts}
\alias{hosts}
\title{Get hosts of XenaHub object}
\usage{
hosts(x)
}
\arguments{
\item{x}{a \link{XenaHub} object}
}
\value{
a character vector contains hosts
}
\description{
Get hosts of XenaHub object
}
\examples{
xe = XenaGenerate(subset = XenaHostNames == "TCGA"); hosts(xe)
}
