% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{showTCGA}
\alias{showTCGA}
\title{Show TCGA data structure by Project ID or ALL}
\usage{
showTCGA(project = "all")
}
\arguments{
\item{project}{a character vector. Can be "all" or one or more of TCGA Project IDs.}
}
\value{
a \code{data.frame} including project data structure information.
}
\description{
This can used to check if data type or file type exist in one or more projects by hand.
}
\examples{
\donttest{
showTCGA("all")
}
}
\seealso{
\code{\link[UCSCXenaTools:availTCGA]{UCSCXenaTools::availTCGA()}}
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
