% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{tcgaAvail}
\alias{tcgaAvail}
\title{TCGA available ProjectID, DataType and FileType}
\usage{
tcgaAvail(which = c("all", "ProjectID", "DataType", "FileType"))
}
\arguments{
\item{which}{a character of \code{c("All", "ProjectID", "DataType", "FileType")}}
}
\description{
TCGA available ProjectID, DataType and FileType
}
\examples{
\donttest{
tcgaAvail("all")
}
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
