% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_xq.R
\name{to_snake}
\alias{to_snake}
\title{Convert camel case to snake case}
\usage{
to_snake(name)
}
\arguments{
\item{name}{a character vector}
}
\value{
same length as \code{name} but with snake case
}
\description{
Convert camel case to snake case
}
\examples{
to_snake("sparseDataRange")
}
