% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\name{UCSCXenaTools-dynamic}
\alias{UCSCXenaTools-dynamic}
\alias{.p_all_cohorts}
\alias{.p_all_datasets}
\alias{.p_all_datasets_n}
\alias{.p_all_field_metadata}
\alias{.p_cohort_samples}
\alias{.p_cohort_summary}
\alias{.p_dataset_fetch}
\alias{.p_dataset_field}
\alias{.p_dataset_field_examples}
\alias{.p_dataset_field_n}
\alias{.p_dataset_gene_probe_avg}
\alias{.p_dataset_gene_probes_values}
\alias{.p_dataset_list}
\alias{.p_dataset_metadata}
\alias{.p_dataset_probe_signature}
\alias{.p_dataset_probe_values}
\alias{.p_dataset_samples}
\alias{.p_dataset_samples_ndense_matrix}
\alias{.p_dataset_sources}
\alias{.p_dataset_status}
\alias{.p_datasets_null_rows}
\alias{.p_feature_list}
\alias{.p_field_codes}
\alias{.p_field_metadata}
\alias{.p_gene_transcripts}
\alias{.p_match_fields}
\alias{.p_probe_count}
\alias{.p_probemap_list}
\alias{.p_ref_gene_exons}
\alias{.p_ref_gene_position}
\alias{.p_ref_gene_range}
\alias{.p_segment_data_examples}
\alias{.p_segmented_data_range}
\alias{.p_sparse_data}
\alias{.p_sparse_data_examples}
\alias{.p_sparse_data_match_field}
\alias{.p_sparse_data_match_field_slow}
\alias{.p_sparse_data_match_partial_field}
\alias{.p_sparse_data_range}
\alias{.p_transcript_expression}
\alias{.xq_all_cohorts}
\alias{.xq_all_datasets}
\alias{.xq_all_datasets_n}
\alias{.xq_all_field_metadata}
\alias{.xq_cohort_samples}
\alias{.xq_cohort_summary}
\alias{.xq_dataset_fetch}
\alias{.xq_dataset_field}
\alias{.xq_dataset_field_examples}
\alias{.xq_dataset_field_n}
\alias{.xq_dataset_gene_probe_avg}
\alias{.xq_dataset_gene_probes_values}
\alias{.xq_dataset_list}
\alias{.xq_dataset_metadata}
\alias{.xq_dataset_probe_signature}
\alias{.xq_dataset_probe_values}
\alias{.xq_dataset_samples}
\alias{.xq_dataset_samples_ndense_matrix}
\alias{.xq_dataset_sources}
\alias{.xq_dataset_status}
\alias{.xq_datasets_null_rows}
\alias{.xq_feature_list}
\alias{.xq_field_codes}
\alias{.xq_field_metadata}
\alias{.xq_gene_transcripts}
\alias{.xq_match_fields}
\alias{.xq_probe_count}
\alias{.xq_probemap_list}
\alias{.xq_ref_gene_exons}
\alias{.xq_ref_gene_position}
\alias{.xq_ref_gene_range}
\alias{.xq_segment_data_examples}
\alias{.xq_segmented_data_range}
\alias{.xq_sparse_data}
\alias{.xq_sparse_data_examples}
\alias{.xq_sparse_data_match_field}
\alias{.xq_sparse_data_match_field_slow}
\alias{.xq_sparse_data_match_partial_field}
\alias{.xq_sparse_data_range}
\alias{.xq_transcript_expression}
\title{UCSC Xena Dynamic Objects}
\format{
Various formats depending on the object (e.g., characters, or functions).
}
\source{
UCSC Xena API
}
\description{
A collection of dynamically generated objects for accessing UCSC Xena data.
These objects are created when the package is loaded via \code{.onAttach}.
}
