% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaHub-class.R
\name{xena_default_hosts}
\alias{xena_default_hosts}
\title{UCSC Xena Default Hosts}
\usage{
xena_default_hosts()
}
\value{
A character vector include current defalut hosts
}
\description{
Return Xena Default hosts
}
\seealso{
\code{\link[UCSCXenaTools]{XenaHub}}
}
\author{
Shixiang Wang
}
