% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaBrowse.R
\name{XenaBrowse}
\alias{XenaBrowse}
\title{View Info of Dataset or Cohort at UCSC Xena Website Using Web browser}
\usage{
XenaBrowse(x, type = c("dataset", "cohort"), multiple = FALSE)
}
\arguments{
\item{x}{a \link{XenaHub} object.}

\item{type}{one of "dataset" and "cohort".}

\item{multiple}{if \code{TRUE}, browse multiple links instead of throwing error.}
}
\description{
This will open dataset/cohort link of UCSC Xena
in user's default browser.
}
\examples{
\donttest{
XenaGenerate(subset = XenaHostNames == "tcgaHub") %>%
  XenaFilter(filterDatasets = "clinical") %>%
  XenaFilter(filterDatasets = "LUAD") -> to_browse
}
}
