% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XenaData}
\alias{XenaData}
\title{Xena Hub Information}
\format{A \code{data.frame}.}
\source{
Generated from UCSC Xena Data Hubs.
}
\description{
This \code{data.frame} is very useful for selecting datasets quickly and indepent on
 API of UCSC Xena Hubs.
}
\examples{
data(XenaData)
str(XenaData)
}
