% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{XenaDownload}
\alias{XenaDownload}
\title{Download datasets form UCSC Xena datasets.}
\usage{
XenaDownload(xquery, destdir = paste0(getwd(), "XenaData"), force = FALSE,
  ...)
}
\arguments{
\item{xquery}{a tibble object generated by \code{XenaQuery} function.}

\item{destdir}{specify a location to store download data. Default is 'XenaData' directory
under work directory.}

\item{force}{logical. if \code{TRUE}, force to download data no matter whether files exist.
Default is \code{FALSE}.}

\item{...}{other argument to \code{download.file} function}
}
\value{
a \code{tibble}
}
\description{
Avaliable datasets list: <https://xenabrowser.net/datapages/>
}
\examples{
xe = XenaHub(hostName="TCGA")
hosts(xe) 
xe_query = XenaQuery(xe)    
# xe_download = XenaDownload(xe_query) # not run
}
