% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaHub-class.R
\docType{class}
\name{XenaHub-class}
\alias{XenaHub-class}
\alias{.XenaHub}
\title{Class XenaHub}
\description{
a S4 class to represent UCSC Xena Data Hubs
}
\section{Slots}{

\describe{
\item{\code{hosts}}{hosts of data hubs}

\item{\code{cohorts}}{cohorts of data hubs}

\item{\code{datasets}}{datasets of data hubs}
}}

