% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaQuery.R
\name{XenaQuery}
\alias{XenaQuery}
\title{Query URL of Datasets before Downloading}
\usage{
XenaQuery(x)
}
\arguments{
\item{x}{a \link{XenaHub} object}
}
\value{
a \code{data.frame} contains hosts, datasets and url
}
\description{
Query URL of Datasets before Downloading
}
\examples{
xe = XenaGenerate(subset = XenaHostNames == "tcgaHub")
hosts(xe)
\dontrun{
xe_query = XenaQuery(xe)
}
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
