% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{XenaQueryProbeMap}
\alias{XenaQueryProbeMap}
\title{Query ProbeMap URL of Datasets}
\usage{
XenaQueryProbeMap(x)
}
\arguments{
\item{x}{a \link{XenaHub} object}
}
\value{
a \code{data.frame} contains hosts, datasets and url
}
\description{
If dataset has no ProbeMap, it will be ignored.
}
\examples{
\donttest{
xe = XenaGenerate(subset = XenaHostNames == "tcgaHub")
hosts(xe)
xe_query = XenaQueryProbeMap(xe)
}
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
