% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaHub-class.R
\name{xena_default_hosts}
\alias{xena_default_hosts}
\title{UCSC Xena Default Hosts}
\usage{
xena_default_hosts()
}
\value{
A character vector include current defalut hosts
}
\description{
Return Xena default hosts
}
\seealso{
\code{\link[UCSCXenaTools:XenaHub]{UCSCXenaTools::XenaHub()}}
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
