% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaHub-class.R
\name{XenaDataUpdate}
\alias{XenaDataUpdate}
\title{Get or Update Newest Data Information of UCSC Xena Data Hubs}
\usage{
XenaDataUpdate(saveTolocal = TRUE)
}
\arguments{
\item{saveTolocal}{logical. Whether save to local R package data directory for permanent use
or Not.}
}
\value{
a \code{data.frame} contains all datasets information of Xena.
}
\description{
Get or Update Newest Data Information of UCSC Xena Data Hubs
}
\examples{

\dontrun{
 XenaDataUpdate() # update newest information to local directory for use
 newest_Xena = XenaDataUpdate(saveTolocal=FALSE) # just get info, not save
}
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
