% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{XenaGenerate}
\alias{XenaGenerate}
\title{Generate and Subset a \code{XenaHub} Object from \code{XenaData}}
\usage{
XenaGenerate(XenaData = UCSCXenaTools::XenaData, subset = TRUE)
}
\arguments{
\item{XenaData}{a data.frame. Default is \code{data(XenaData)}. The input of this
argument can only be \code{data(XenaData)} or its subset.}

\item{subset}{logical expression indicating elements or rows to keep.}
}
\value{
a \code{XenaHub} object.
}
\description{
Generate and Subset a \code{XenaHub} Object from \code{XenaData}
}
\examples{
# 1 get all datasets
XenaGenerate()
# 2
XenaGenerate(subset = XenaCohorts == "TCGA Breast Cancer (BRCA)")
# 3
XenaGenerate(subset = grepl("BRCA", XenaCohorts))
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
