% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenaHub-class.R
\docType{class}
\name{XenaHub-class}
\alias{XenaHub-class}
\alias{.XenaHub}
\title{@title a S4 class to represent UCSC Xena Data Hubs
@slot hosts hosts of data hubs
@slot cohorts cohorts of data hubs
@slot datasets datasets of data hubs
@importFrom methods new}
\description{
@title a S4 class to represent UCSC Xena Data Hubs
@slot hosts hosts of data hubs
@slot cohorts cohorts of data hubs
@slot datasets datasets of data hubs
@importFrom methods new
}
