% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{XenaQuery}
\alias{XenaQuery}
\title{Query url of datasets before download}
\usage{
XenaQuery(x)
}
\arguments{
\item{x}{a \code{XenaHub} object}
}
\value{
a \code{tibble} contains hosts, datasets and url
}
\description{
Query url of datasets before download
}
\examples{
xe = XenaGenerate(subset = XenaHostNames == "TCGA")
hosts(xe)
xe_query = XenaQuery(xe)
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
